requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorlover==0.3.0
cufflinks==0.17.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
django-cors-headers==3.5.0
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
entrypoints==0.3
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
optional-django==0.3.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.11.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
retrying==1.3.3
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
webpack==6.0.0
widgetsnbextension==3.5.1
